Change to buffered output - does not run properly
For some reason, I was using unbuffered output. Changed this to buffered
to work with other platforms without having to resort to conditional
includes and such.

For now, this program crashes when parsing the input file, and I can't
tell how it was supposed to work originally. Instead of trying to get
this to run as-is, I'll spend future time upgrading this to modern C++.